+++ src/UriInfoTest.php
@@ -215,4 +215,35 @@ final class UriInfoTest extends TestCase
             ],
         ];
     }
+
+    /**
+     * @dataProvider getCrossOriginExamples
+     */
+    public function testIsCrossOrigin(string $original, string $modified, bool $expected): void
+    {
+        self::assertSame($expected, UriInfo::isCrossOrigin(Uri::createFromString($original), Http::createFromString($modified)));
+    }
+
+    /**
+     * @return array<string, array{0:string, 1:string, 2:bool}>
+     */
+    public function getCrossOriginExamples(): array
+    {
+        return [
+            'different path' => ['http://example.com/123', 'http://example.com/', false],
+            'same port with default value (1)' => ['https://example.com/123', 'https://example.com:443/', false],
+            'same port with default value (2)' => ['ws://example.com:80/123', 'ws://example.com/', false],
+            'same explicit port' => ['wss://example.com:443/123', 'wss://example.com:443/', false],
+            'same origin with i18n host' => ['https://xn--bb-bjab.be./path', 'https://Bébé.BE./path', false],
+            'same origin using a blob' => ['blob:https://mozilla.org:443/', 'https://mozilla.org/123', false],
+            'different scheme' => ['https://example.com/123', 'ftp://example.com/', true],
+            'different host' => ['ftp://example.com/123', 'ftp://www.example.com/123', true],
+            'different port implicit' => ['https://example.com/123', 'https://example.com:81/', true],
+            'different port explicit' => ['https://example.com:80/123', 'https://example.com:81/', true],
+            'same scheme different port' => ['https://example.com:443/123', 'https://example.com:444/', true],
+            'comparing two opaque URI' => ['ldap://ldap.example.net', 'ldap://ldap.example.net', true],
+            'comparing a URI with an origin and one with an opaque origin' => ['https://example.com:443/123', 'ldap://ldap.example.net', true],
+            'cross origin using a blob' => ['blob:http://mozilla.org:443/', 'https://mozilla.org/123', true],
+        ];
+    }
 }